Fix the bug where compilation failed for those who didn't use Timing.
The compile script now compiles only the bare library without the
support for the timing app. To compile the library for timing, use the
-t option: ./compile.sh -t Library

+++ compile.sh
@@ -1,5 +1,6 @@
 #!/bin/bash
 
+timing=0
 scriptfolder="$HOME/Library/Application Scripts/com.omnigroup.OmniFocus3"
 usage="$(basename "$0") [-h] DIRECTORY
 
@@ -7,8 +8,10 @@ where:
 	-h         show this help and exit
 	DIRECTORY  directory of the script"
 
-while getopts ':h' option; do
+while getopts ':ht' option; do
 	case "$option" in
+		t)  timing=1
+				;;
 		h) 	echo "$usage"
 				exit
 				;;
@@ -18,19 +21,34 @@ while getopts ':h' option; do
 				;;
 	esac
 done
+shift $((OPTIND-1))
 
 input=${1////}
 
 if [ "$input" == "Library" ]; then
-	src="Library/omnifocuslib.applescript"
-	trgt="$HOME/Library/Script Libraries/omnifocuslib.scpt"
+	srcone="Library/omnifocuslib.applescript"
+	trgtone="$HOME/Library/Script Libraries/omnifocuslib.scpt"
+	if [ -f "$trgtone" ]; then
+		rm "$trgtone"
+	fi
+	if [ $timing == 1 ]; then
+		srctwo="Library/timinglib.applescript"
+		trgttwo="$HOME/Library/Script Libraries/timinglib.scpt"
+		if [ -f "$trgttwo" ]; then
+			rm "$trgttwo"
+		fi
+		cmd="osacompile -o \"$trgttwo\" \"$srctwo\""
+	else
+		cmd="printf \"\""
+	fi
+	osacompile -o "$trgtone" "$srcone" && eval "$cmd" && echo "DONE"
 else
 	lowercase=$(echo "$input" | tr '[:upper:]' '[:lower:]')
 	bn=${lowercase//[[:blank:]]/}
 	src="$input"/"$bn".applescript
 	trgt="$scriptfolder"/"$bn".scpt
+	if [ -f "$trgt" ]; then
+		rm "$trgt"
+	fi
+	osacompile -o "$trgt" "$src" && echo "DONE"
 fi
-
-rm "$trgt"
-
-osacompile -o "$trgt" "$src" && echo "DONE"
